requirements.txt
flask==3.1.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
cachetools==5.5.2
python-dotenv==1.1.0
pytest==8.3.5
pytest-mock==3.14.0
